feat: [main] 커스텀 데코레이터를 만들어 포스트 삭제 및 수정 시 id를 이용해 판별

+++ src/routes/board/board.controller.ts
@@ -15,6 +15,7 @@ import {
 import { CreateBoardDto } from './dto/create-board.dto';
 import { UpdateBoardDto } from './dto/update-board.dto';
 import { JwtAuthGuard } from 'src/auth/jwt-auth.guard';
+import { UserInfo } from 'src/decorators/user-info.decorator';
 
 @Controller('board')
 @ApiTags('board')
@@ -33,20 +34,34 @@ export class BoardController {
 
   @UseGuards(JwtAuthGuard)
   @Post()
-  create(@Body(new ValidationPipe()) data: CreateBoardDto) {
-    return this.boardService.create(data);
+  create(
+    @UserInfo() userInfo,
+    @Body(new ValidationPipe()) data: CreateBoardDto,
+  ) {
+    return this.boardService.create({ userId: userInfo.id, ...data });
   }
 
+  @UseGuards(JwtAuthGuard)
   @Put(':id')
-  update(
+  async update(
+    @UserInfo() userInfo,
     @Body(new ValidationPipe()) data: UpdateBoardDto,
     @Param('id', ParseIntPipe) id: number,
   ) {
+    const board = await this.boardService.find(id);
+    if (userInfo.id !== board.id) {
+      return '주인이 아닙니다.';
+    }
     return this.boardService.update(data, id);
   }
 
+  @UseGuards(JwtAuthGuard)
   @Delete(':id')
-  remove(@Param('id', ParseIntPipe) id: number) {
+  async remove(@UserInfo() userInfo, @Param('id', ParseIntPipe) id: number) {
+    const board = await this.boardService.find(id);
+    if (userInfo.id !== board.id) {
+      return '주인이 아닙니다.';
+    }
     return this.boardService.delete(id);
   }
 }
